need to make sure only author or admin can delete comments

+++ finalproject/index.php
@@ -1,6 +1,10 @@
 <?php
 //Create Session
 session_start();
+
+// Check if user is admin
+$is_admin = $_SESSION['user_id'] == 1;
+
 //header
 $page_title = "Barry's Blog";
 include('header.php');
@@ -20,10 +24,11 @@ if (isset($_SESSION['first_name'])) {
 	echo "You currently logged in as {$_SESSION['first_name']}. Welcome to our website!";
 }
 
+
 //***********************************************
 //DELETE LOGIC START
 //***********************************************
-if (isset($_GET['delete_id'])) {
+if (isset($_GET['delete_id']) && $is_admin) {
 	$delete_id = mysqli_real_escape_string($dbc, trim($_GET['delete_id']));
 	$delete_query = "DELETE from blogposts WHERE blogpost_id = $delete_id";
 	$delete_results = mysqli_query($dbc, $delete_query);
@@ -39,8 +44,14 @@ if (isset($_GET['delete_id'])) {
 //UPDATE LOGIC START
 //***********************************************
 
-if (isset($_GET['blogpost_updated']) && $_GET['blogpost_updated'] == 1) {
+// Check if a blogpost was updated
+if (isset($_GET['blogpost_updated']) && $_GET['blogpost_updated'] == 1 && $is_admin) {
 	$notifications[] = array('alert-level' => 'success', 'message' => 'Blog post successfully updated');
+}
+
+// Check if a blogpost was created
+if (isset($_GET['blogpost_inserted']) && $_GET['blogpost_inserted'] == 1 && $is_admin) {
+	$notifications[] = array('alert-level' => 'success', 'message' => 'Blog post successfully inserted');
 }
 
 function get_update_modal($update_id, $update_title, $update_body, $errors = array()) {
@@ -292,18 +303,21 @@ $select_all_results = mysqli_query($dbc, $select_all_query);
 									</p>
 									<div class="d-flex justify-content-between align-items-center">
 										<p class="card-text mb-0"><small class="text-muted">Last updated <?php echo $date ?></small></p>
-										<div>
-											<a class="btn btn-outline-primary" href="update_blogpost.php?update_id=<?php echo $id ?>">
-												<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-pencil-fill" viewBox="0 0 16 16">
-													<path d="M12.854.146a.5.5 0 0 0-.707 0L10.5 1.793 14.207 5.5l1.647-1.646a.5.5 0 0 0 0-.708l-3-3zm.646 6.061L9.793 2.5 3.293 9H3.5a.5.5 0 0 1 .5.5v.5h.5a.5.5 0 0 1 .5.5v.5h.5a.5.5 0 0 1 .5.5v.5h.5a.5.5 0 0 1 .5.5v.207l6.5-6.5zm-7.468 7.468A.5.5 0 0 1 6 13.5V13h-.5a.5.5 0 0 1-.5-.5V12h-.5a.5.5 0 0 1-.5-.5V11h-.5a.5.5 0 0 1-.5-.5V10h-.5a.499.499 0 0 1-.175-.032l-.179.178a.5.5 0 0 0-.11.168l-2 5a.5.5 0 0 0 .65.65l5-2a.5.5 0 0 0 .168-.11l.178-.178z" />
-												</svg>
-											</a>
-											<a class="btn btn-outline-danger" href="index.php?<?php echo $page_location ?>&delete_id=<?php echo $id; ?>">
-												<svg xmlns=" http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-trash3-fill" viewBox="0 0 16 16">
-													<path d="M11 1.5v1h3.5a.5.5 0 0 1 0 1h-.538l-.853 10.66A2 2 0 0 1 11.115 16h-6.23a2 2 0 0 1-1.994-1.84L2.038 3.5H1.5a.5.5 0 0 1 0-1H5v-1A1.5 1.5 0 0 1 6.5 0h3A1.5 1.5 0 0 1 11 1.5Zm-5 0v1h4v-1a.5.5 0 0 0-.5-.5h-3a.5.5 0 0 0-.5.5ZM4.5 5.029l.5 8.5a.5.5 0 1 0 .998-.06l-.5-8.5a.5.5 0 1 0-.998.06Zm6.53-.528a.5.5 0 0 0-.528.47l-.5 8.5a.5.5 0 0 0 .998.058l.5-8.5a.5.5 0 0 0-.47-.528ZM8 4.5a.5.5 0 0 0-.5.5v8.5a.5.5 0 0 0 1 0V5a.5.5 0 0 0-.5-.5Z" />
-												</svg>
-											</a>
-										</div>
+										<!-- Only show update and delete buttons if user is admin -->
+										<?php if ($is_admin) : ?>
+											<div>
+												<a class="btn btn-outline-primary" href="update_blogpost.php?update_id=<?php echo $id ?>">
+													<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-pencil-fill" viewBox="0 0 16 16">
+														<path d="M12.854.146a.5.5 0 0 0-.707 0L10.5 1.793 14.207 5.5l1.647-1.646a.5.5 0 0 0 0-.708l-3-3zm.646 6.061L9.793 2.5 3.293 9H3.5a.5.5 0 0 1 .5.5v.5h.5a.5.5 0 0 1 .5.5v.5h.5a.5.5 0 0 1 .5.5v.5h.5a.5.5 0 0 1 .5.5v.207l6.5-6.5zm-7.468 7.468A.5.5 0 0 1 6 13.5V13h-.5a.5.5 0 0 1-.5-.5V12h-.5a.5.5 0 0 1-.5-.5V11h-.5a.5.5 0 0 1-.5-.5V10h-.5a.499.499 0 0 1-.175-.032l-.179.178a.5.5 0 0 0-.11.168l-2 5a.5.5 0 0 0 .65.65l5-2a.5.5 0 0 0 .168-.11l.178-.178z" />
+													</svg>
+												</a>
+												<a class="btn btn-outline-danger" href="index.php?<?php echo $page_location ?>&delete_id=<?php echo $id; ?>">
+													<svg xmlns=" http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="bi bi-trash3-fill" viewBox="0 0 16 16">
+														<path d="M11 1.5v1h3.5a.5.5 0 0 1 0 1h-.538l-.853 10.66A2 2 0 0 1 11.115 16h-6.23a2 2 0 0 1-1.994-1.84L2.038 3.5H1.5a.5.5 0 0 1 0-1H5v-1A1.5 1.5 0 0 1 6.5 0h3A1.5 1.5 0 0 1 11 1.5Zm-5 0v1h4v-1a.5.5 0 0 0-.5-.5h-3a.5.5 0 0 0-.5.5ZM4.5 5.029l.5 8.5a.5.5 0 1 0 .998-.06l-.5-8.5a.5.5 0 1 0-.998.06Zm6.53-.528a.5.5 0 0 0-.528.47l-.5 8.5a.5.5 0 0 0 .998.058l.5-8.5a.5.5 0 0 0-.47-.528ZM8 4.5a.5.5 0 0 0-.5.5v8.5a.5.5 0 0 0 1 0V5a.5.5 0 0 0-.5-.5Z" />
+													</svg>
+												</a>
+											</div>
+										<?php endif; ?>
 									</div>
 								</div>
 							</div>
